Migrate all cost data processing to `process_cost_data.py` (#1719)
* Create process_cost_data.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix merge conflict

* Add Yerbol suggestion, add type hinting to process_cost_data

* update some script docstring related to costs because its no longer in params

* pre-commit run --all

---------
