+++ scripts/augmented_line_connections.py
@@ -35,7 +35,6 @@ import numpy as np
 import pandas as pd
 import pypsa
 from _helpers import configure_logging, create_logger
-from add_electricity import load_costs
 from networkx.algorithms import complement
 from networkx.algorithms.connectivity.edge_augmentation import k_edge_augmentation
 from pypsa.geo import haversine_pts
@@ -73,12 +72,7 @@ if __name__ == "__main__":
         raise SystemExit(0)
 
     Nyears = n.snapshot_weightings.sum().values[0] / 8760.0
-    costs = load_costs(
-        snakemake.input.tech_costs,
-        snakemake.params.costs,
-        snakemake.params.electricity,
-        Nyears,
-    )
+    costs = pd.read_csv(snakemake.input.tech_costs, index_col=0)
     # TODO: Implement below comment in future. Requires transformer consideration.
     # component_type = {"False": "Line", "True":  "Link"}.get(snakemake.params.hvdc_as_lines)
     options = snakemake.params.augmented_line_connection
