+++ scripts/make_summary.py
@@ -11,17 +11,11 @@ Relevant Settings
 -----------------
 .. code:: yaml
 
-    costs:
-        USD2013_to_EUR2013:
-        discountrate:
-        marginal_cost:
-        capital_cost:
-    electricity:
-        max_hours:
+    scenario:
 
 .. seealso::
     Documentation of the configuration file ``config.yaml`` at
-    :ref:`costs_cf`, :ref:`electricity_cf`
+    :ref:`toplevel_cf`
 
 Inputs
 ------
@@ -56,7 +50,7 @@ import os
 import pandas as pd
 import pypsa
 from _helpers import configure_logging
-from add_electricity import create_logger, load_costs, update_transmission_costs
+from add_electricity import create_logger, update_transmission_costs
 
 idx = pd.IndexSlice
 
@@ -484,7 +478,7 @@ outputs = [
 ]
 
 
-def make_summaries(networks_dict, inputs, cost_config, elec_config, country="all"):
+def make_summaries(networks_dict, inputs, country="all"):
     columns = pd.MultiIndex.from_tuples(
         networks_dict.keys(), names=["simpl", "clusters", "ll", "opts"]
     )
@@ -509,13 +503,7 @@ def make_summaries(networks_dict, inputs, cost_config, elec_config, country="all
         if country != "all":
             n = n[n.buses.country == country]
 
-        Nyears = n.snapshot_weightings.objective.sum() / 8760.0
-        costs = load_costs(
-            inputs.tech_costs,
-            cost_config,
-            elec_config,
-            Nyears,
-        )
+        costs = pd.read_csv(inputs.tech_costs, index_col=0)
         update_transmission_costs(n, costs, simple_hvdc_costs=False)
 
         assign_carriers(n)
@@ -580,8 +568,6 @@ if __name__ == "__main__":
     dfs = make_summaries(
         networks_dict,
         snakemake.input,
-        snakemake.params.costs,
-        snakemake.params.electricity,
         country=snakemake.wildcards.country,
     )
